+++ src/main/java/com/aetherteam/aether/item/combat/loot/ValkyrieLanceItem.java
@@ -1,6 +1,5 @@
 package com.aetherteam.aether.item.combat.loot;
 
-import com.aetherteam.aether.item.AetherItems;
 import com.aetherteam.aether.item.combat.AetherItemTiers;
 import com.aetherteam.aether.item.tools.abilities.ValkyrieTool;
 import net.minecraft.world.entity.EquipmentSlotGroup;
@@ -10,13 +9,13 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.SwordItem;
 
 public class ValkyrieLanceItem extends SwordItem implements ValkyrieTool {
-    public ValkyrieLanceItem() {
+    public ValkyrieLanceItem(Item.Properties properties) {
         super(
                 AetherItemTiers.VALKYRIE,
                 SwordItem.createAttributes(AetherItemTiers.VALKYRIE, 3, -2.7F)
                         .withModifierAdded(Attributes.BLOCK_INTERACTION_RANGE, new AttributeModifier(BLOCK_INTERACTION_RANGE_MODIFIER_UUID, RANGE_MODIFER, AttributeModifier.Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND)
                         .withModifierAdded(Attributes.ENTITY_INTERACTION_RANGE, new AttributeModifier(ENTITY_INTERACTION_RANGE_MODIFIER_UUID, RANGE_MODIFER, AttributeModifier.Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND),
-                new Item.Properties().rarity(AetherItems.AETHER_LOOT)
+                properties
         );
     }